refactor: move add unit facade method and helpers to addunit file
