+++ apiserver/facades/client/application/application.go
@@ -1012,76 +1012,6 @@ func (api *APIBase) Unexpose(ctx context.Context, args params.ApplicationUnexpos
 	return nil
 }
 
-// AddUnits adds a given number of units to an application.
-func (api *APIBase) AddUnits(ctx context.Context, args params.AddApplicationUnits) (params.AddApplicationUnitsResults, error) {
-	if api.modelType == model.CAAS {
-		return params.AddApplicationUnitsResults{}, errors.NotSupportedf("adding units to a container-based model")
-	}
-
-	if err := api.checkCanWrite(ctx); err != nil {
-		return params.AddApplicationUnitsResults{}, errors.Trace(err)
-	}
-	if err := api.check.ChangeAllowed(ctx); err != nil {
-		return params.AddApplicationUnitsResults{}, errors.Trace(err)
-	}
-
-	units, err := api.addApplicationUnits(ctx, args)
-	if err != nil {
-		return params.AddApplicationUnitsResults{}, errors.Trace(err)
-	}
-	return params.AddApplicationUnitsResults{
-		Units: transform.Slice(units, coreunit.Name.String),
-	}, nil
-}
-
-// addApplicationUnits adds a given number of units to an application.
-func (api *APIBase) addApplicationUnits(
-	ctx context.Context, args params.AddApplicationUnits,
-) ([]coreunit.Name, error) {
-	if args.NumUnits < 1 {
-		return nil, errors.New("must add at least one unit")
-	}
-
-	if api.modelType == model.CAAS {
-		// In a CAAS model, there are no machines for
-		// units to be assigned to.
-		if len(args.AttachStorage) > 0 {
-			return nil, errors.Errorf(
-				"AttachStorage may not be specified for %s models",
-				api.modelType,
-			)
-		}
-		if len(args.Placement) > 1 {
-			return nil, errors.Errorf(
-				"only 1 placement directive is supported for %s models, got %d",
-				api.modelType,
-				len(args.Placement),
-			)
-		}
-	}
-
-	// Parse storage tags in AttachStorage.
-	if len(args.AttachStorage) > 0 && args.NumUnits != 1 {
-		return nil, errors.Errorf("AttachStorage is non-empty, but NumUnits is %d", args.NumUnits)
-	}
-	attachStorage := make([]names.StorageTag, len(args.AttachStorage))
-	for i, tagString := range args.AttachStorage {
-		tag, err := names.ParseStorageTag(tagString)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		attachStorage[i] = tag
-	}
-
-	return api.addUnits(
-		ctx,
-		args.ApplicationName,
-		args.NumUnits,
-		args.Placement,
-		attachStorage,
-	)
-}
-
 // DestroyUnit removes a given set of application units.
 func (api *APIBase) DestroyUnit(ctx context.Context, args params.DestroyUnitsParams) (params.DestroyUnitResults, error) {
 	if api.modelType == model.CAAS {
